Simple restructure, added noset to handle standalone `Lekvar.set` method calls
